chore: move navigation helper to existing file
